test: move user profile tests out of library
